Major application architecture overhaul
This commit represents a significant overhaul of the application's architecture, focusing on enhancing the MVVM pattern, improving database interactions, and refining run management and image processing tasks. Key changes include the removal of redundant variables and classes, such as the `once` variable in `JsonViewer.xaml.cs` and entire `Controller` class in `Controller.cs`, indicating a significant refactoring effort. New functionalities have been added, such as the introduction of a `Controller` class in the `LabelVal.Run` namespace, marked as `partial` and inheriting from `ObservableObject`, which includes a delegate, an event for run state changes, and various properties and methods related to database operations and run logic. The addition of a `RunStates` enum and updates to properties in `ImageEntry.cs` to include `[ObservableProperty]` and `[property: SQLite.Ignore]` attributes suggest enhancements to property change notifications and database interaction handling. Furthermore, adjustments in ViewModel properties and the simplification of the database opening process in `RunDatabase.cs` streamline database interactions and improve data encapsulation. The removal of files such as `Result.cs` and `Run.cs`, alongside modifications in `RunControl.cs` and `SimulatorFileHandler.cs`, reflect a significant refactoring or redesign of the application's run management logic. Lastly, the use of modern C# features such as nullable reference types and the Community Toolkit's MVVM features in the `Controller` implementation and WebSocketEvents adjustments in the `V5_REST_Lib` namespace indicate an effort to adhere to current best practices in software development.